Remove old gulp and bower files and set up builds using rollup instead. Remove lodash dependency
