--- bower.json
@@ -1,39 +1,0 @@
-{
-  "name": "vue-isotope",
-  "authors": [
-    "desmaisons_david <desmaisons_david@hotmail.com>"
-  ],
-  "description": "vue isotope directive",
-  "main": "dist/vue_isotope.js",
-  "keywords": [
-    "vue",
-    "list",
-    "isotope"
-  ],
-  "license": "MIT",
-  "homepage": "",
-  "ignore": [
-    "**/.*",
-    "node_modules",
-    "bower_components",
-    "test",
-    "tests",
-    "examples",
-    "*.json",
-    "*.md",
-    "gulpfile.js"
-  ],
-  "dependencies": {
-    "vue": "^2.0.1",
-    "isotope": "^3.0.1",
-    "lodash": "^4.13.1"
-  },
-  "devDependencies": {
-    "isotope-cells-by-column": "^1.1.3",
-    "isotope-packery": "^2.0.0",
-    "isotope-cells-by-row": "^1.1.3",
-    "isotope-masonry-horizontal": "^2.0.0",
-    "isotope-horizontal": "^2.0.0",
-    "isotope-fit-columns": "^1.1.3"
-  }
-}
